+++ pkg/regexutils/regex.go
@@ -13,13 +13,31 @@ type WholeStringMatcher interface {
 	MatchWholeString(s string) bool
 }
 
+// Flags represents regex flags.
+// See the Flags table in https://github.com/google/re2/wiki/Syntax for the list
+// of regex flags.
+// For simplicity, we have not implemented all flags here, only what
+// we have used, but we are using a struct to enable seamlessly adding more.
+type Flags struct {
+	CaseInsensitive bool
+}
+
+// regexRepr represents these flags in the regexp format.
+// See the Flags table in https://github.com/google/re2/wiki/Syntax.
+func (f *Flags) regexRepr() string {
+	if f.CaseInsensitive {
+		return "i"
+	}
+	return ""
+}
+
 // CompileWholeStringMatcher takes a regex and compiles it into a WholeStringMatcher.
 // An empty regex matches _all_ strings.
-func CompileWholeStringMatcher(re string) (WholeStringMatcher, error) {
+func CompileWholeStringMatcher(re string, flags Flags) (WholeStringMatcher, error) {
 	if re == "" {
 		return &wholeStringMatcher{}, nil
 	}
-	compiled, err := regexp.Compile(fmt.Sprintf("^(?:%s)$", re))
+	compiled, err := regexp.Compile(fmt.Sprintf("^(?%s:%s)$", flags.regexRepr(), re))
 	if err != nil {
 		return nil, errors.Wrap(err, "invalid regex")
 	}